pyproject.toml
[project]
name = "cascade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Kanyang Ying", email = "lexugeyky@outlook.com" }]
requires-python = ">=3.12"
dependencies = [
    "common-expression-language>=0.5.0",
    "croniter>=6.0.0",
    "numpy>=2.3.2",
    "ortools>=9.14.6206",
    "prompt-toolkit>=3.0.51",
    "pydantic>=2.11.7",
    "python-slugify>=8.0.4",
    "pytimeparse>=1.1.8",
    "pyyaml>=6.0.2",
    "scipy>=1.16.1",
]

[project.scripts]
cascade = "cascade:main"

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"
